+++ src/main.rs
@@ -14,6 +14,7 @@ mod deps;
 mod emoji;
 mod images;
 mod model;
+mod notify;
 mod plugins;
 mod tree;